// typescript/models/CreateOrModifyPolicyResponse.ts
/**
 * Metronome API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0.0
 * Contact: info@privatedataservices.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class CreateOrModifyPolicyResponse {
    'policyId': string;
    'fieldsUpdated': Array<string>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "policyId",
            "baseName": "policy-id",
            "type": "string",
            "format": ""
        },
        {
            "name": "fieldsUpdated",
            "baseName": "fields-updated",
            "type": "Array<string>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CreateOrModifyPolicyResponse.attributeTypeMap;
    }

    public constructor() {
    }
}