// typescript/models/ExpireSubItemResponse.ts
/**
 * Metronome API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0.0
 * Contact: info@privatedataservices.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class ExpireSubItemResponse {
    'itemId': string;
    'subItemId': string;
    'fieldsUpdated': Array<string>;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "itemId",
            "baseName": "item-id",
            "type": "string",
            "format": ""
        },
        {
            "name": "subItemId",
            "baseName": "sub-item-id",
            "type": "string",
            "format": ""
        },
        {
            "name": "fieldsUpdated",
            "baseName": "fields-updated",
            "type": "Array<string>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ExpireSubItemResponse.attributeTypeMap;
    }

    public constructor() {
    }
}